legal_rag_api/requirements.txt
django==5.2
djangorestframework==3.16.0
django-cors-headers==4.3.1
beautifulsoup4==4.12.2
requests==2.31.0
openai>=1.3.0
faiss-cpu>=1.7.4
numpy>=1.24.0
python-dotenv>=1.0.0
langchain>=0.3.0
langchain-openai>=0.0.1
scikit-learn>=1.0.0
pillow>=9.0.0
pdf2image>=1.16.0
pytesseract>=0.3.10
pdfminer.six>=20221105
pymupdf>=1.21.0
tiktoken>=0.4.0
sentence-transformers>=2.2.2 